【2017-1-13 18:00】Sms页面及WifiKey页面初步提交

+++ As_SmartLinkV3.1/app/src/main/java/com/alcatel/smartlinkv3/ui/activity/UsageSettingActivity.java
@@ -682,7 +682,7 @@ public class UsageSettingActivity extends BaseActivity implements OnClickListene
 			UsageSettingModel statistic = BusinessMannager.getInstance()
 					.getUsageSettings();
 			
-			m_consumptionValue.setText(CommonUtil.ConvertTrafficToStringFromMB(this, (long)statistic.HUsedData));
+			m_consumptionValue.setText(CommonUtil.ConvertTrafficToStringFromMB(this, statistic.HUsedData));
 
 			m_monthlyValue.setEnabled(true);
 			m_monthlyVal = statistic.HMonthlyPlan;
